+++ generate-vapid-key.js
@@ -1,0 +1,4 @@
+const webpush = require('web-push');
+
+const vapidKeys = webpush.generateVAPIDKeys();
+console.log(JSON.stringify(vapidKeys));