+++ models/ride.mjs
@@ -4,12 +4,15 @@ import Luxon from 'luxon';
 import nanoid from 'nanoid';
 import stateMachinePlugin from '@rentspree/mongoose-state-machine';
 import gliphone from 'google-libphonenumber';
-import stateMachine, { DRAFTED } from './status';
+import { CAN_ACCESS_OWN_DATA_ON_RIDE, CAN_ACCESS_PERSONAL_DATA_ON_RIDE } from './rights';
+import stateMachine, { DRAFTED, DELIVERED } from './status';
 import config from '../services/config';
 import { sendSMS } from '../services/twilio';
 import createdAtPlugin from './helpers/created-at';
+import cleanObjectPlugin from './helpers/object-cleaner';
 
 const DEFAULT_TIMEZONE = config.get('default_timezone');
+const MODEL_NAME = 'Ride';
 const { DateTime, Duration } = Luxon;
 const { PhoneNumberFormat, PhoneNumberUtil } = gliphone;
 const { Schema, Types } = mongoose;
@@ -35,10 +38,22 @@ const RideSchema = new Schema({
   },
   end: Date,
   owner: {
-    _id: { type: mongoose.Types.ObjectId, alias: 'owner.id' },
-    firstname: String,
-    lastname: String,
-    email: String,
+    _id: {
+      type: mongoose.Types.ObjectId,
+      alias: 'owner.id',
+    },
+    firstname: {
+      type: String,
+      canEmit: [CAN_ACCESS_PERSONAL_DATA_ON_RIDE, CAN_ACCESS_OWN_DATA_ON_RIDE],
+    },
+    lastname: {
+      type: String,
+      canEmit: [CAN_ACCESS_PERSONAL_DATA_ON_RIDE, CAN_ACCESS_OWN_DATA_ON_RIDE],
+    },
+    email: {
+      type: String,
+      canEmit: [CAN_ACCESS_PERSONAL_DATA_ON_RIDE, CAN_ACCESS_OWN_DATA_ON_RIDE],
+    },
   },
   departure: {
     _id: { type: String, required: true, alias: 'departure.id' },
@@ -91,7 +106,10 @@ const RideSchema = new Schema({
     },
   },
   comments: String,
-  userComments: String,
+  userComments: {
+    type: String,
+    canEmit: [CAN_ACCESS_PERSONAL_DATA_ON_RIDE, CAN_ACCESS_OWN_DATA_ON_RIDE],
+  },
   passengersCount: {
     type: Number,
     default: 1,
@@ -104,6 +122,7 @@ const RideSchema = new Schema({
 });
 
 RideSchema.plugin(createdAtPlugin);
+RideSchema.plugin(cleanObjectPlugin, MODEL_NAME);
 RideSchema.plugin(stateMachinePlugin.default, { stateMachine });
 
 RideSchema.pre('validate', async function beforeSave() {
@@ -157,9 +176,45 @@ RideSchema.statics.castId = (v) => {
   }
 };
 
-RideSchema.statics.filtersWithin = function filtersWithin(start, end, f = {}) {
-  const filters = f;
-  filters.$or = [
+RideSchema.statics.formatFilters = function formatFilters(rawFilters, queryFilter) {
+  let filter = {
+    ...rawFilters,
+    ...queryFilter,
+    ...this.filtersWithin(queryFilter.start, queryFilter.end),
+  };
+
+  delete filter.start;
+  delete filter.end;
+
+
+  if (filter.current) {
+    let status;
+    if (filter.current === 'false') {
+      status = { status: DELIVERED };
+    } else {
+      status = { $nor: [{ status: DELIVERED }] };
+    }
+
+    filter = {
+      ...filter,
+      ...status,
+    };
+
+    delete filter.current;
+  }
+
+
+  if (!filter) {
+    return null;
+  }
+  return filter;
+};
+
+RideSchema.statics.filtersWithin = function filtersWithin(rawStart, rawEnd) {
+  const queryFilter = {};
+  const start = new Date(rawStart);
+  const end = new Date(rawEnd);
+  queryFilter.$or = [
     {
       start: {
         $lte: start,
@@ -197,21 +252,15 @@ RideSchema.statics.filtersWithin = function filtersWithin(start, end, f = {}) {
       },
     },
   ];
-  return filters;
+  return queryFilter;
 };
 
-RideSchema.statics.findWithin = function findWithin(start, end, filters = {}, ...rest) {
-  return this.find(
-    this.filtersWithin(start, end, filters),
-    ...rest,
-  );
+RideSchema.statics.findWithin = function findWithin(...params) {
+  return this.find(this.formatFilters(...params));
 };
 
-RideSchema.statics.countDocumentsWithin = function countDocumentsWithin(start, end, filters = {}, ...rest) {
-  return this.countDocuments(
-    this.filtersWithin(start, end, filters),
-    ...rest,
-  );
+RideSchema.statics.countDocumentsWithin = function countDocumentsWithin(...params) {
+  return this.countDocuments(this.formatFilters(...params));
 };
 
 RideSchema.methods.findDriverPosition = async function findDriverPosition() {
@@ -259,4 +308,4 @@ RideSchema.methods.getSatisfactionQuestionnaireURL = function getSatisfactionQue
   return `${config.get('satisfaction_questionnaire_url')}`;
 };
 
-export default mongoose.model('Ride', RideSchema);
+export default mongoose.model(MODEL_NAME, RideSchema);